+++ src/components/query-list/query-list.js
@@ -7,7 +7,7 @@ import PrevNextButtons from '../prev-next-buttons';
 import ImpaginationReplacement from './ImpaginationReplacement';
 import { PAGE_SIZE } from '../../constants';
 
-export default class QueryList extends Component {
+class QueryList extends Component {
   static propTypes = {
     collection: PropTypes.object.isRequired,
     fetch: PropTypes.func.isRequired,
@@ -17,7 +17,7 @@ export default class QueryList extends Component {
     length: PropTypes.number,
     notFoundMessage: PropTypes.oneOfType([
       PropTypes.string,
-      PropTypes.node
+      PropTypes.node,
     ]).isRequired,
     offset: PropTypes.number,
     onUpdateOffset: PropTypes.func.isRequired,
@@ -41,6 +41,8 @@ export default class QueryList extends Component {
     this.state = {
       offset: this.props.offset || 0,
     };
+
+    this.listFirstItem = null;
   }
 
   updateOffset = (offset) => {
@@ -59,6 +61,16 @@ export default class QueryList extends Component {
     }
   }
 
+  setListFirstItemRef = (el) => {
+    this.listFirstItem = el;
+  };
+
+  focusListFirstItem = () => {
+    if (this.listFirstItem) {
+      this.listFirstItem.focus();
+    }
+  };
+
   getPrevNextButtons = () => {
     const {
       collection: {
@@ -76,11 +88,23 @@ export default class QueryList extends Component {
           totalResults={length}
           fetch={this.updatePage}
           page={page}
+          setFocus={this.focusListFirstItem}
         />
       );
     }
 
     return null;
+  }
+
+  defineIsListFirstItem = ({
+    item,
+    state,
+  }) => {
+    const [firstRecord] = state.records;
+    const recordId = firstRecord?.id;
+    const itemId = item.content?.id;
+
+    return itemId === recordId;
   }
 
   render() {
@@ -104,6 +128,41 @@ export default class QueryList extends Component {
 
     const offsetProp = isMainPageSearch ? page : offset;
 
+    const getContent = (state) => {
+      if (!state.length) {
+        return notFoundMessage;
+      }
+
+      return (
+        <ScrollView
+          items={state}
+          length={length}
+          offset={offsetProp}
+          isMainPageSearch
+          itemHeight={itemHeight}
+          scrollable={scrollable}
+          onUpdate={this.updateOffset}
+          queryListName={type}
+          fullWidth={fullWidth}
+          prevNextButtons={this.getPrevNextButtons()}
+        >
+          {item => (
+            <div
+              ref={this.defineIsListFirstItem({ item, state })
+                ? (el) => this.setListFirstItemRef(el)
+                : null
+              }
+              tabIndex={this.defineIsListFirstItem({ item, state }) ? 0 : null} // eslint-disable-line jsx-a11y/no-noninteractive-tabindex
+              className={styles['list-item']}
+              data-test-query-list-item
+            >
+              {renderItem(item)}
+            </div>
+          )}
+        </ScrollView>
+      );
+    };
+
     return (
       <ImpaginationReplacement
         pageSize={pageSize}
@@ -112,34 +171,20 @@ export default class QueryList extends Component {
         fetch={fetch}
       >
         {state => (
-          state.hasRejected && !state.length ? (
-            <div className={styles.error} data-test-query-list-error={type}>
-              {state.rejected[0].error[0].title}
-            </div>
-          ) : !state.length
-            ? notFoundMessage
-            : (
-              <ScrollView
-                items={state}
-                length={length}
-                offset={offsetProp}
-                isMainPageSearch
-                itemHeight={itemHeight}
-                scrollable={scrollable}
-                onUpdate={this.updateOffset}
-                queryListName={type}
-                fullWidth={fullWidth}
-                prevNextButtons={this.getPrevNextButtons()}
+          state.hasRejected && !state.length
+            ? (
+              <div
+                className={styles.error}
+                data-test-query-list-error={type}
               >
-                {item => (
-                  <div className={styles['list-item']} data-test-query-list-item>
-                    {renderItem(item)}
-                  </div>
-                )}
-              </ScrollView>
+                {state.rejected[0].error[0].title}
+              </div>
             )
+            : getContent(state)
         )}
       </ImpaginationReplacement>
     );
   }
 }
+
+export default QueryList;